requirements.txt
annotated-types==0.7.0
arabic-reshaper==3.0.0
asgiref==3.9.1
asn1crypto==1.5.1
blis==1.3.0
catalogue==2.0.10
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
cloudpathlib==0.21.1
cmudict==1.0.33
colorama==0.4.6
confection==0.1.5
cryptography==45.0.5
cssselect2==0.8.0
cymem==2.0.11
defusedxml==0.7.1
Django==5.2.4
django-widget-tweaks==1.5.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
gunicorn==23.0.0
html5lib==1.1
idna==3.10
importlib_metadata==8.7.0
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
langcodes==3.5.0
language_data==1.3.0
language_tool_python==2.9.4
lxml==6.0.0
marisa-trie==1.2.1
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
murmurhash==1.0.13
mysqlclient==2.2.7
numpy==2.3.1
oauthlib==3.3.1
oscrypto==1.3.0
packaging==25.0
pdf2image==1.17.0
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
preshed==3.0.10
psutil==7.0.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyHanko==0.29.1
pyhanko-certvalidator==0.27.0
PyJWT==2.10.1
pypdf==5.7.0
pypdfium2==4.30.1
pyphen==0.17.2
pytesseract==0.3.13
python-bidi==0.6.6
python-docx==1.2.0
python-dotenv==1.1.1
python3-openid==3.2.0
PyYAML==6.0.2
reportlab==4.4.2
requests==2.32.4
requests-oauthlib==2.0.0
rich==14.0.0
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smart_open==7.3.0.post1
social-auth-app-django==5.5.1
social-auth-core==4.7.0
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.3
srsly==2.5.1
svglib==1.5.1
textstat==0.7.7
thinc==8.3.6
threadpoolctl==3.6.0
tinycss2==1.4.0
toml==0.10.2
tqdm==4.67.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritools==5.0.0
urllib3==2.5.0
wasabi==1.1.3
weasel==0.4.1
webencodings==0.5.1
wrapt==1.17.2
xhtml2pdf==0.2.17
zipp==3.23.0
